+++ NEWS.md
@@ -1,3 +1,9 @@
+# imuf 0.4.0
+* Added `rotV()`to rotate a 3-vector by a quaternion
+* Added `walking_shin_1` dataset
+* Better function documentation
+* Added a vignette
+
 # imuf 0.3.0
 
 # imuf 0.2.1